refactor: enhance home page
